libschedutil: Consistently add NULL check
Problem: The ops alloc_cb checked for a NULL util pointer but the
other callbacks cancel_cb and free_cb did not.

Solution: Consistently check for NULL in all callbacks, but use an
assert check instead.

Remove an unnecessary NULL check that returns EINVAL.  Instead replace
it was an assert check.  Add assert check consistently in other callbacks
in ops.